dont use eval for audio processing js code, make new js file instead
